cu29_units

Module si

Modules

absement

Absement (base unit meter second, m · s).

acceleration

Acceleration (base unit meter per second squared, m · s⁻²).

action

Action (base unit joule second, kg ⋅ m² ⋅ s⁻¹).

amount_of_substance

Amount of substance (base unit mole, mol).

angle

Angle (dimensionless quantity).

angular_absement

Angular Absement (base unit radian second, s).

angular_acceleration

Angular acceleration (base unit radian per second squared, s⁻²).

angular_jerk

Angular jerk (base unit radian per second cubed, s⁻³).

angular_momentum

Angular momentum (base unit kilogram square meter per second, kg · m² · s⁻¹).

angular_velocity

Angular velocity (base unit radian per second, s⁻¹).

area

Area (base unit square meter, m²).

areal_density_of_states

Areal density of states (base unit 1 / square meter joule, kg⁻¹ · m⁻⁴ · s²).

areal_heat_capacity

Areal heat capacity (base unit joule per square meter kelvin, kg · s⁻² · K⁻¹).

areal_mass_density

Areal mass density (base unit kilogram per square meter, m⁻² · kg).

areal_number_density

Areal number density (base unit 1 per square meter, m⁻²).

areal_number_rate

Areal number rate (base unit 1 per square meter second, m⁻² · s⁻¹).

available_energy

Available energy (base unit joule per kilogram, m² · s⁻²).

capacitance

Capacitance (base unit farad, m⁻² · kg⁻¹ · s⁴ · A²).

catalytic_activity

Catalytic activity (base unit katal, mol · s⁻¹).

catalytic_activity_concentration

Catalytic activity concentration (base unit katal per cubic meter, mol · s⁻¹ · m⁻³).

curvature

Curvature (base unit radian per meter, m⁻¹).

diffusion_coefficient

Diffusion coefficient (base unit square meter per second, m² · s⁻¹).

dynamic_viscosity

Dynamic viscosity (base unit pascal second, kg · m⁻¹ · s⁻¹).

electric_charge

Electric charge (base unit coulomb, A · s).

electric_charge_areal_density

Electric charge areal density (base unit coulomb per square meter, m⁻² · A · s).

electric_charge_linear_density

Electric charge linear density (base unit coulomb per meter, m⁻¹ · A · s).

electric_charge_volumetric_density

Electric charge volumetric density (base unit coulomb per cubic meter, m⁻³ · A · s).

electric_current

Electric current (base unit ampere, A).

electric_current_density

Electric current density (base unit ampere per square meter, m⁻² · A).

electric_dipole_moment

Electric dipole moment (base unit coulomb meter, m · s · A).

electric_displacement_field

Electric displacement field (base unit coulomb per square meter, m⁻² · A · s).

electric_field

Electric field (base unit volt per meter, m ⋅ kg ⋅ s⁻³ ⋅ A⁻¹).

electric_flux

Electric Flux (base unit volt meter, m³ ⋅ kg ⋅ s⁻³ ⋅ A⁻¹).

electric_permittivity

Electric permittivity (base unit farad per meter, m⁻³ · kg⁻¹ · s⁴ · A²).

electric_potential

Electric potential (base unit volt, m² · kg · s⁻³ · A⁻¹).

electric_quadrupole_moment

Electric quadrupole moment (base unit coulomb square meter, m² · s · A).

electrical_conductance

Electrical conductance (base unit siemens, m⁻² · kg⁻¹ · s³ · A²).

electrical_conductivity

Electrical conductivity (base unit siemens per meter, m⁻³ · kg⁻¹ · s³ · A²).

electrical_mobility

Electrical mobility (base unit square meter per volt second, kg⁻¹ · s² · A).

electrical_resistance

Electrical resistance (base unit ohm, m² · kg · s⁻³ · A⁻²).

electrical_resistivity

Electrical resistivity (base unit ohm meter, m³ · kg · s⁻³ · A⁻²).

energy

Energy (base unit joule, kg · m² · s⁻²).

f32

f64

force

Force (base unit newton, kg · m · s⁻²).

frequency

Frequency (base unit hertz, s⁻¹).

frequency_drift

Frequency drift (base unit hertz per second, s⁻²).

heat_capacity

Heat capacity (base unit joule per kelvin, kg · m² · s⁻² · K⁻¹).

heat_flux_density

Heat flux density (base unit watt per square meter, kg · s⁻³).

heat_transfer

Heat transfer (base unit watt per square meter kelvin, kg · s⁻³ · K⁻¹).

inductance

Inductance (base unit henry, m² · kg · s⁻² · A⁻²).

information

Information (dimensionless quantity).

information_rate

Information rate (base unit byte per second, s⁻¹).

inverse_velocity

Inverse velocity (base unit second per meter, s · m⁻¹).

jerk

Jerk (base unit meter per second cubed, m · s⁻³).

kinematic_viscosity

Kinematic viscosity (base unit square meter per second, m² · s⁻¹).

length

Length (base unit meter, m).

linear_density_of_states

Linear density of states (base unit 1 / meter joule, kg⁻¹ · m⁻³ · s²).

linear_mass_density

Linear mass density (base unit kilogram per meter, m⁻¹ · kg).

linear_number_density

Linear number density (base unit 1 per meter, m⁻¹).

linear_number_rate

Linear number rate (base unit 1 per meter second, m⁻¹ · s⁻¹).

linear_power_density

Linear power density (base unit watt per meter, m · kg · s⁻³).

luminance

Luminance (base unit candela per square meter, cd · m⁻²).

luminous_intensity

Luminous intensity (base unit candela, cd).

magnetic_field_strength

Magnetic field strength (H field) (base unit ampere per meter, m⁻¹ · A).

magnetic_flux

Magnetic flux (base unit weber, m² · kg · s⁻² · A⁻¹).

magnetic_flux_density

Magnetic flux density (base unit tesla, kg · s⁻² · A⁻¹).

magnetic_moment

Magnetic moment (base unit ampere square meter A · m²).

magnetic_permeability

Magnetic permeability (base unit henry per meter, m · kg · s⁻² · A⁻²).

mass

Mass (base unit kilogram, kg).

mass_concentration

Mass concentration (base unit kilogram per cubic meter, kg · m⁻³).

mass_density

Mass density (base unit kilogram per cubic meter, kg · m⁻³).

mass_flux

Mass flux (base unit kilogram per square meter second, m⁻² · kg · s⁻¹).

mass_per_energy

Mass per energy (base unit kilogram per joule, m⁻² · s²).

mass_rate

Mass rate (base unit kilogram per second, kg · s⁻¹).

molality

Molality (base unit mole per kilogram, kg⁻¹ · mol).

molar_concentration

Molar concentration (base unit mole per cubic meter, mol · m⁻³).

molar_energy

Molar energy (base unit joule per mole, kg · m² · s⁻² · mol⁻¹).

molar_flux

Molar flux (base unit mole per square meter second, m⁻² · s⁻¹ · mol).

molar_heat_capacity

Molar heat capacity (base unit joule per kelvin mole, m² · kg · s⁻² · K⁻¹ · mol⁻¹).

molar_mass

Molar mass (base unit kilogram per mole, kg · mol⁻¹).

molar_radioactivity

Molar radioactivity (base unit becquerel per mole, s⁻¹ · mol⁻¹).

molar_volume

Molar volume (base unit cubic meter per mole, m³ · mol⁻¹).

moment_of_inertia

Moment of inertia (base unit kilogram square meter, kg · m² ).

momentum

Momentum (base unit kilogram meter per second, kg · m · s⁻¹).

power

Power (base unit watt, m² · kg · s⁻³).

power_rate

Power rate (base unit watt per second, m² · kg · s⁻⁴).

pressure

Pressure (base unit pascal, kg · m⁻¹ · s⁻²).

radiant_exposure

Radiant exposure (base unit joule per square meter, kg · s⁻²).

radioactivity

Radioactivity (base unit becquerel, s⁻¹).

ratio

Ratio (dimensionless quantity).

reciprocal_length

Reciprocal length (base unit reciprocal meter, m⁻¹).

solid_angle

Solid angle (dimensionless quantity).

specific_area

Specific area (base unit square meter per kilogram, m² · kg⁻¹).

specific_heat_capacity

Specific heat capacity (base unit joule per kilogram kelvin, m² · s⁻² · K⁻¹).

specific_power

Specific power (base unit watt per kilogram, m² · s⁻³).

specific_radioactivity

Specific radioactivity (base unit becquerel per kilogram, kg⁻¹ · s⁻¹).

specific_volume

Specific volume (base unit cubic meter per kilogram, m³ · kg⁻¹).

surface_electric_current_density

Surface electric current density (base unit ampere per meter, ).

surface_tension

Surface tension (base unit newton per meter, kg · s⁻²).

temperature_coefficient

Temperature coefficient (base unit 1 / kelvin, K⁻¹).

temperature_gradient

Temperature gradient (base unit kelvin per meter, m⁻¹ · K).

temperature_interval

Temperature interval (base unit kelvin, K).

thermal_conductance

Thermal conductance (base unit watt per kelvin, kg · m² · s⁻³ · K⁻¹).

thermal_conductivity

Thermal conductivity (base unit watt per meter kelvin, kg · m · s⁻³ · K⁻¹).

thermal_resistance

Thermal resistance (base unit kelvin per watt, kg⁻¹ · m⁻² · s³ · K).

thermodynamic_temperature

Thermodynamic temperature (base unit kelvin, K).

time

Time (base unit second, s).

torque

Torque (aka moment of force) (base unit newton meter, kg · m² · s⁻²).

velocity

Velocity (base unit meter per second, m · s⁻¹).

volume

Volume (base unit cubic meter, m³).

volume_rate

Volume rate (base unit cubic meter per second, m³ · s⁻¹).

volumetric_density_of_states

Volumetric density of states (base unit 1 / cubic meter joule, kg⁻¹ · m⁻⁵ · s²).

volumetric_heat_capacity

Volumetric heat capacity (base unit joule per cubic meter kelvin, m⁻¹ · kg · s⁻² · K⁻¹).

volumetric_number_density

Volumetric number density (base unit 1 per cubic meter, m⁻³).

volumetric_number_rate

Volumetric number rate (base unit 1 per cubic meter second, m⁻³ · s⁻¹).

volumetric_power_density

Volumetric power density (base unit watt per cubic meter, m⁻¹ · kg · s⁻³).

Type Aliases

ISQ

International System of Quantities (ISQ).

SI

International System of Units (SI).