cu29_units::si

Module absement

Absement (base unit meter second, m · s).

Structs

attometer_second

attometer seconds

centimeter_second

centimeter seconds

decameter_second

decameter seconds

decimeter_second

decimeter seconds

exameter_second

exameter seconds

femtometer_second

femtometer seconds

foot_second

foot seconds

gigameter_second

gigameter seconds

hectometer_second

hectometer seconds

inch_second

inch seconds

kilometer_hour

kilometer hours

kilometer_second

kilometer seconds

megameter_second

megameter seconds

meter_second

meter seconds

micrometer_second

micrometer seconds

millimeter_second

millimeter seconds

nanometer_second

nanometer seconds

petameter_second

petameter seconds

picometer_second

picometer seconds

terameter_second

terameter seconds

yoctometer_second

yoctometer seconds

yottameter_second

yottameter seconds

zeptometer_second

zeptometer seconds

zettameter_second

zettameter seconds

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Absement

Absement (base unit meter second, m · s).

Dimension

Dimension of absement, LT (base unit meter second, m · s).