cu29_units::si

Module acceleration

Acceleration (base unit meter per second squared, m · s⁻²).

Structs

attometer_per_second_squared

attometers per second squared

centimeter_per_second_squared

centimeters per second squared

decameter_per_second_squared

decameters per second squared

decimeter_per_second_squared

decimeters per second squared

exameter_per_second_squared

exameters per second squared

femtometer_per_second_squared

femtometers per second squared

foot_per_second_squared

feet per second squared

galileo

galileos

gigameter_per_second_squared

gigameters per second squared

hectometer_per_second_squared

hectometers per second squared

inch_per_second_squared

inches per second squared

kilometer_per_second_squared

kilometers per second squared

megameter_per_second_squared

megameters per second squared

meter_per_second_squared

meters per second squared

micrometer_per_second_squared

micrometers per second squared

millimeter_per_minute_squared

millimeters per minute squared

millimeter_per_second_squared

millimeters per second squared

nanometer_per_second_squared

nanometers per second squared

petameter_per_second_squared

petameters per second squared

picometer_per_second_squared

picometers per second squared

standard_gravity

standard accelerations of gravity

terameter_per_second_squared

terameters per second squared

yoctometer_per_second_squared

yoctometers per second squared

yottameter_per_second_squared

yottameters per second squared

zeptometer_per_second_squared

zeptometers per second squared

zettameter_per_second_squared

zettameters per second squared

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Acceleration

Acceleration (base unit meter per second squared, m · s⁻²).

Dimension

Dimension of acceleration, LT⁻² (base unit meter per second squared, m · s⁻²).