cu29_units::si

Module amount_of_substance

Amount of substance (base unit mole, mol).

Structs

attomole

attomoles

centimole

centimoles

decamole

decamoles

decimole

decimoles

examole

examoles

femtomole

femtomoles

gigamole

gigamoles

hectomole

hectomoles

kilomole

kilomoles

megamole

megamoles

micromole

micromoles

millimole

millimoles

mole

The mole is the SI unit of amount of substance. One mole contains exactly 6.022 140 76 × 10²³ elementary entities. This number is the fixed numerical value of the Avogadro constant, NA, when expressed in the unit mol⁻¹ and is called the Avogadro number.The amount of substance, symbol n, of a system is a measure of the number of specified elementary entities. An elementary entity may be an atom, a molecule, an ion, an electron, any other particle or specified group of particles.

nanomole

nanomoles

particle

One elementary entity may be an atom, a molecule, an ion, an electron, any other particle or specified group of particles.

petamole

petamoles

picomole

picomoles

standard_centimeter

Amount of ideal gas contained in a volume of cubic centimeter at standard temperature (O°C) and pressure (1 bar)

standard_cubic_foot

Amount of ideal gas contained in a volume of cubic foot at standard temperature (O°C) and pressure (1 bar)

standard_cubic_meter

Amount of ideal gas contained in a volume of cubic meter at standard temperature (O°C) and pressure (1 bar)

standard_liter

Amount of ideal gas contained in a volume of liter at standard temperature (O°C) and pressure (1 bar)

teramole

teramoles

yoctomole

yoctomoles

yottamole

yottamoles

zeptomole

zeptomoles

zettamole

zettamoles

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

AmountOfSubstance

Amount of substance (base unit mole, mol).

Dimension

Dimension of amount of substance, N (base unit mole, mol).