cu29_units::si

Module angular_acceleration

Angular acceleration (base unit radian per second squared, s⁻²).

Structs

degree_per_second_squared

degrees per second squared

radian_per_second_squared

Derived unit of angular acceleration.

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

AngularAcceleration

Angular acceleration (base unit radian per second squared, s⁻²).

Dimension

Dimension of angular acceleration, T⁻² (base unit radian per second squared, s⁻²).