cu29_units::si

Module angular_momentum

Angular momentum (base unit kilogram square meter per second, kg · m² · s⁻¹).

Structs

attonewton_meter_second

attonewton meter seconds

centinewton_meter_second

centinewton meter seconds

decanewton_meter_second

decanewton meter seconds

decinewton_meter_second

decinewton meter seconds

exanewton_meter_second

exanewton meter seconds

femtonewton_meter_second

femtonewton meter seconds

giganewton_meter_second

giganewton meter seconds

hectonewton_meter_second

hectonewton meter seconds

kilogram_square_meter_per_second

Derived unit of angular velocity.

kilonewton_meter_second

kilonewton meter seconds

meganewton_meter_second

meganewton meter seconds

micronewton_meter_second

micronewton meter seconds

millinewton_meter_second

millinewton meter seconds

nanonewton_meter_second

nanonewton meter seconds

newton_attometer_second

newton attometer seconds

newton_centimeter_second

newton centimeter seconds

newton_decameter_second

newton decameter seconds

newton_decimeter_second

newton decimeter seconds

newton_exameter_second

newton exameter seconds

newton_femtometer_second

newton femtometer seconds

newton_gigameter_second

newton gigameter seconds

newton_hectometer_second

newton hectometer seconds

newton_kilometer_second

newton kilometer seconds

newton_megameter_second

newton megameter seconds

newton_meter_second

newton meter seconds

newton_micrometer_second

newton micrometer seconds

newton_millimeter_second

newton millimeter seconds

newton_nanometer_second

newton nanometer seconds

newton_petameter_second

newton petameter seconds

newton_picometer_second

newton picometer seconds

newton_terameter_second

newton terameter seconds

newton_yoctometer_second

newton yoctometer seconds

newton_yottameter_second

newton yottameter seconds

newton_zeptometer_second

newton zeptometer seconds

newton_zettameter_second

newton zettameter seconds

petanewton_meter_second

petanewton meter seconds

piconewton_meter_second

piconewton meter seconds

teranewton_meter_second

teranewton meter seconds

yoctonewton_meter_second

yoctonewton meter seconds

yottanewton_meter_second

yottanewton meter seconds

zeptonewton_meter_second

zeptonewton meter seconds

zettanewton_meter_second

zettanewton meter seconds

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

AngularMomentum

Angular momentum (base unit kilogram square meter per second, kg · m² · s⁻¹).

Dimension

Dimension of angular momentum, L²MT⁻¹ (base unit kilogram square meter per second, kg · m² · s⁻¹).