cu29_units::si::angular_momentum

Type Alias AngularMomentum

pub type AngularMomentum<U, V> = Quantity<dyn Dimension<T = NInt<UInt<UTerm, B1>>, N = Z0, J = Z0, L = PInt<UInt<UInt<UTerm, B1>, B0>>, Kind = dyn AngleKind, I = Z0, Th = Z0, M = PInt<UInt<UTerm, B1>>>, U, V>;

Angular momentum (base unit kilogram square meter per second, kg · m² · s⁻¹).

Generic Parameters

U: Base units.

V: Underlying storage type.

Aliased Type

#[repr(transparent)]
pub struct AngularMomentum<U, V> {
    pub dimension: PhantomData<dyn Dimension<T = NInt<UInt<UTerm, B1>>, N = Z0, J = Z0, L = PInt<UInt<UInt<UTerm, B1>, B0>>, Kind = dyn AngleKind, I = Z0, Th = Z0, M = PInt<UInt<UTerm, B1>>>>,
    pub units: PhantomData<U>,
    pub value: V,
}

Fields
dimension: PhantomData<dyn Dimension<T = NInt<UInt<UTerm, B1>>, N = Z0, J = Z0, L = PInt<UInt<UInt<UTerm, B1>, B0>>, Kind = dyn AngleKind, I = Z0, Th = Z0, M = PInt<UInt<UTerm, B1>>>>

Quantity dimension. See [Dimension].

units: PhantomData<U>

Quantity base units. See [Units].

value: V

Quantity value stored in the base units for the quantity.