cu29_units::si

Module area

Area (base unit square meter, m²).

Structs

acre

acres

are

ares

barn

barns

circular_mil

circular mils

hectare

hectares

square_attometer

square attometers

square_centimeter

square centimeters

square_decameter

square decameters

square_decimeter

square decimeters

square_exameter

square exameters

square_femtometer

square femtometers

square_foot

square feet

square_gigameter

square gigameters

square_hectometer

square hectometers

square_inch

square inches

square_kilometer

square kilometers

square_megameter

square megameters

square_meter

square meters

square_micrometer

square micrometers

square_mile

square miles

square_millimeter

square millimeters

square_nanometer

square nanometers

square_petameter

square petameters

square_picometer

square picometers

square_terameter

square terameters

square_yard

square yards

square_yoctometer

square yoctometers

square_yottameter

square yottameters

square_zeptometer

square zeptometers

square_zettameter

square zettameters

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Area

Area (base unit square meter, m²).

Dimension

Dimension of area, L² (base unit square meter, m²).