cu29_units::si

Module available_energy

Available energy (base unit joule per kilogram, m² · s⁻²).

Structs

attojoule_per_kilogram

attojoules per kilogram

btu_it_per_pound

British thermal units (IT) per pound

btu_per_pound

British thermal units per pound

calorie_it_per_gram

calories (IT) per gram

calorie_per_gram

calories per gram

centijoule_per_kilogram

centijoules per kilogram

decajoule_per_kilogram

decajoules per kilogram

decijoule_per_kilogram

decijoules per kilogram

exajoule_per_kilogram

exajoules per kilogram

femtojoule_per_kilogram

femtojoules per kilogram

gigajoule_per_kilogram

gigajoules per kilogram

hectojoule_per_kilogram

hectojoules per kilogram

joule_per_attogram

joules per attogram

joule_per_exagram

joules per exagram

joule_per_femtogram

joules per femtogram

joule_per_gigagram

joules per gigagram

joule_per_gram

joules per gram

joule_per_kilogram

Derived unit of available energy.

joule_per_megagram

joules per megagram

joule_per_microgram

joules per microgram

joule_per_milligram

joules per milligram

joule_per_nanogram

joules per nanogram

joule_per_petagram

joules per petagram

joule_per_picogram

joules per picogram

joule_per_teragram

joules per teragram

joule_per_yottagram

joules per yottagram

joule_per_zeptogram

joules per zeptogram

joule_per_zettagram

joules per zettagram

kilojoule_per_kilogram

kilojoules per kilogram

megajoule_per_kilogram

megajoules per kilogram

microjoule_per_kilogram

microjoules per kilogram

millijoule_per_kilogram

millijoules per kilogram

nanojoule_per_kilogram

nanojoules per kilogram

petajoule_per_kilogram

petajoules per kilogram

picojoule_per_kilogram

picojoules per kilogram

terajoule_per_kilogram

terajoules per kilogram

yoctojoule_per_kilogram

yoctojoules per kilogram

yottajoule_per_kilogram

yottajoules per kilogram

zeptojoule_per_kilogram

zeptojoules per kilogram

zettajoule_per_kilogram

zettajoules per kilogram

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

AvailableEnergy

Available energy (base unit joule per kilogram, m² · s⁻²).

Dimension

Dimension of available energy, L²T⁻² (base unit joule per kilogram, m² · s⁻²).