cu29_units::si

Module capacitance

Capacitance (base unit farad, m⁻² · kg⁻¹ · s⁴ · A²).

Structs

abfarad

abfarads

attofarad

attofarads

centifarad

centifarads

decafarad

decafarads

decifarad

decifarads

exafarad

exafarads

farad

Derived unit of capacitance.

femtofarad

femtofarads

gigafarad

gigafarads

hectofarad

hectofarads

kilofarad

kilofarads

megafarad

megafarads

microfarad

microfarads

millifarad

millifarads

nanofarad

nanofarads

petafarad

petafarads

picofarad

picofarads

statfarad

statfarads

terafarad

terafarads

yoctofarad

yoctofarads

yottafarad

yottafarads

zeptofarad

zeptofarads

zettafarad

zettafarads

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Capacitance

Capacitance (base unit farad, m⁻² · kg⁻¹ · s⁴ · A²).

Dimension

Dimension of capacitance, L⁻²M⁻¹T⁴I² (base unit farad, m⁻² · kg⁻¹ · s⁴ · A²).