cu29_units::si

Module catalytic_activity

Catalytic activity (base unit katal, mol · s⁻¹).

Structs

atto_enzyme_unit

atto enzyme units

attokatal

attokatals

centi_enzyme_unit

centi enzyme units

centikatal

centikatals

deca_enzyme_unit

deca enzyme units

decakatal

decakatals

deci_enzyme_unit

deci enzyme units

decikatal

decikatals

enzyme_unit

enzyme units

exa_enzyme_unit

exa enzyme units

exakatal

exakatals

femto_enzyme_unit

femto enzyme units

femtokatal

femtokatals

giga_enzyme_unit

giga enzyme units

gigakatal

gigakatals

hecto_enzyme_unit

hecto enzyme units

hectokatal

hectokatals

katal

katals

kilo_enzyme_unit

kilo enzyme units

kilokatal

kilokatals

mega_enzyme_unit

mega enzyme units

megakatal

megakatals

micro_enzyme_unit

micro enzyme units

microkatal

microkatals

milli_enzyme_unit

milli enzyme units

millikatal

millikatals

mole_per_second

moles per second

nano_enzyme_unit

nano enzyme units

nanokatal

nanokatals

particle_per_second

particles per second

peta_enzyme_unit

peta enzyme units

petakatal

petakatals

pico_enzyme_unit

pico enzyme units

picokatal

picokatals

standard_centimeter_per_minute

standard centimeters per minute

standard_cubic_foot_per_minute

standard cubic feet per hour

standard_cubic_meter_per_minute

standard cubic meters per minute

standard_liter_per_minute

standard liters per minute

tera_enzyme_unit

tera enzyme units

terakatal

terakatals

yocto_enzyme_unit

yocto enzyme units

yoctokatal

yoctokatals

yotta_enzyme_unit

yotta enzyme units

yottakatal

yottakatals

zepto_enzyme_unit

zepto enzyme units

zeptokatal

zeptokatals

zetta_enzyme_unit

zetta enzyme units

zettakatal

zettakatals

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

CatalyticActivity

Catalytic activity (base unit katal, mol · s⁻¹).

Dimension

Dimension of catalytic activity, T⁻¹N (base unit katal, mol · s⁻¹).