cu29_units::si

Module catalytic_activity_concentration

Catalytic activity concentration (base unit katal per cubic meter, mol · s⁻¹ · m⁻³).

Structs

atto_enzyme_unit_per_cubic_meter

atto enzyme units per cubic meter

attokatal_per_cubic_meter

attokatals per cubic meter

centi_enzyme_unit_per_cubic_meter

centi enzyme units per cubic meter

centikatal_per_cubic_meter

centikatals per cubic meter

deca_enzyme_unit_per_cubic_meter

deca enzyme units per cubic meter

decakatal_per_cubic_meter

decakatals per cubic meter

deci_enzyme_unit_per_cubic_meter

deci enzyme units per cubic meter

decikatal_per_cubic_meter

decikatals per cubic meter

enzyme_unit_per_cubic_decimeter

enzyme units per cubic decimeter

enzyme_unit_per_cubic_meter

enzyme units per cubic meter

enzyme_unit_per_deciliter

enzyme units per deciliter

enzyme_unit_per_liter

enzyme units per liter

enzyme_unit_per_milliliter

enzyme units per milliliter

exa_enzyme_unit_per_cubic_meter

exa enzyme units per cubic meter

exakatal_per_cubic_meter

exakatals per cubic meter

femto_enzyme_unit_per_cubic_decimeter

femto enzyme units per cubic decimeter

femto_enzyme_unit_per_cubic_meter

femto enzyme units per cubic meter

femto_enzyme_unit_per_deciliter

femto enzyme units per deciliter

femto_enzyme_unit_per_liter

femto enzyme units per liter

femto_enzyme_unit_per_milliliter

femto enzyme units per milliliter

femtokatal_per_cubic_decimeter

femtokatals per cubic decimeter

femtokatal_per_cubic_meter

femtokatals per cubic meter

femtokatal_per_deciliter

femtokatals per deciliter

femtokatal_per_liter

femtokatals per liter

femtokatal_per_milliliter

femtokatals per milliliter

giga_enzyme_unit_per_cubic_meter

giga enzyme units per cubic meter

gigakatal_per_cubic_meter

gigakatals per cubic meter

hecto_enzyme_unit_per_cubic_meter

hecto enzyme units per cubic meter

hectokatal_per_cubic_meter

hectokatals per cubic meter

katal_per_cubic_decimeter

katals per cubic decimeter

katal_per_cubic_meter

katals per cubic meter

katal_per_deciliter

katals per deciliter

katal_per_liter

katals per liter

katal_per_milliliter

katals per milliliter

kilo_enzyme_unit_per_cubic_decimeter

kilo enzyme units per cubic decimeter

kilo_enzyme_unit_per_cubic_meter

kilo enzyme units per cubic meter

kilo_enzyme_unit_per_deciliter

kilo enzyme units per deciliter

kilo_enzyme_unit_per_liter

kilo enzyme units per liter

kilo_enzyme_unit_per_milliliter

kilo enzyme units per milliliter

kilokatal_per_cubic_decimeter

kilokatals per cubic decimeter

kilokatal_per_cubic_meter

kilokatals per cubic meter

kilokatal_per_deciliter

kilokatals per deciliter

kilokatal_per_liter

kilokatals per liter

kilokatal_per_milliliter

kilokatals per milliliter

mega_enzyme_unit_per_cubic_meter

mega enzyme units per cubic meter

megakatal_per_cubic_meter

megakatals per cubic meter

micro_enzyme_unit_per_cubic_decimeter

micro enzyme units per cubic decimeter

micro_enzyme_unit_per_cubic_meter

micro enzyme units per cubic meter

micro_enzyme_unit_per_deciliter

micro enzyme units per deciliter

micro_enzyme_unit_per_liter

micro enzyme units per liter

micro_enzyme_unit_per_milliliter

micro enzyme units per milliliter

microkatal_per_cubic_decimeter

microkatals per cubic decimeter

microkatal_per_cubic_meter

microkatals per cubic meter

microkatal_per_deciliter

microkatals per deciliter

microkatal_per_liter

microkatals per liter

microkatal_per_milliliter

microkatals per milliliter

milli_enzyme_unit_per_cubic_decimeter

milli enzyme units per cubic decimeter

milli_enzyme_unit_per_cubic_meter

milli enzyme units per cubic meter

milli_enzyme_unit_per_deciliter

milli enzyme units per deciliter

milli_enzyme_unit_per_liter

milli enzyme units per liter

milli_enzyme_unit_per_milliliter

milli enzyme units per milliliter

millikatal_per_cubic_decimeter

millikatals per cubic decimeter

millikatal_per_cubic_meter

millikatals per cubic meter

millikatal_per_deciliter

millikatals per deciliter

millikatal_per_liter

millikatals per liter

millikatal_per_milliliter

millikatals per milliliter

nano_enzyme_unit_per_cubic_decimeter

nano enzyme units per cubic decimeter

nano_enzyme_unit_per_cubic_meter

nano enzyme units per cubic meter

nano_enzyme_unit_per_deciliter

nano enzyme units per deciliter

nano_enzyme_unit_per_liter

nano enzyme units per liter

nano_enzyme_unit_per_milliliter

nano enzyme units per milliliter

nanokatal_per_cubic_decimeter

nanokatals per cubic decimeter

nanokatal_per_cubic_meter

nanokatals per cubic meter

nanokatal_per_deciliter

nanokatals per deciliter

nanokatal_per_liter

nanokatals per liter

nanokatal_per_milliliter

nanokatals per milliliter

particle_per_second_cubic_meter

particles per second cubic meter

peta_enzyme_unit_per_cubic_meter

peta enzyme units per cubic meter

petakatal_per_cubic_meter

petakatals per cubic meter

pico_enzyme_unit_per_cubic_decimeter

pico enzyme units per cubic decimeter

pico_enzyme_unit_per_cubic_meter

pico enzyme units per cubic meter

pico_enzyme_unit_per_deciliter

pico enzyme units per deciliter

pico_enzyme_unit_per_liter

pico enzyme units per liter

pico_enzyme_unit_per_milliliter

pico enzyme units per milliliter

picokatal_per_cubic_decimeter

picokatals per cubic decimeter

picokatal_per_cubic_meter

picokatals per cubic meter

picokatal_per_deciliter

picokatals per deciliter

picokatal_per_liter

picokatals per liter

picokatal_per_milliliter

picokatals per milliliter

tera_enzyme_unit_per_cubic_meter

tera enzyme units per cubic meter

terakatal_per_cubic_meter

terakatals per cubic meter

yocto_enzyme_unit_per_cubic_meter

yocto enzyme units per cubic meter

yoctokatal_per_cubic_meter

yoctokatals per cubic meter

yotta_enzyme_unit_per_cubic_meter

yotta enzyme units per cubic meter

yottakatal_per_cubic_meter

yottakatals per cubic meter

zepto_enzyme_unit_per_cubic_meter

zepto enzyme units per cubic meter

zeptokatal_per_cubic_meter

zeptokatals per cubic meter

zetta_enzyme_unit_per_cubic_meter

zetta enzyme units per cubic meter

zettakatal_per_cubic_meter

zettakatals per cubic meter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

CatalyticActivityConcentration

Catalytic activity concentration (base unit katal per cubic meter, mol · s⁻¹ · m⁻³).

Dimension

Dimension of catalytic activity concentration, L⁻³T⁻¹N (base unit katal per cubic meter, mol · s⁻¹ · m⁻³).