cu29_units::si

Module dynamic_viscosity

Dynamic viscosity (base unit pascal second, kg · m⁻¹ · s⁻¹).

Structs

centipoise

centipoises

gram_per_centimeter_second

grams per centimeter second

micropascal_second

micropascal seconds

millipascal_second

millipascal seconds

pascal_second

pascal seconds

poise

poises

pound_force_second_per_square_foot

pound-force seconds per square foot

pound_force_second_per_square_inch

pound-force seconds per square inch

pound_per_foot_hour

pounds per foot hour

pound_per_foot_second

pounds per foot second

reyn

reyns

slug_per_foot_second

slugs per foot second

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of dynamic viscosity, L⁻¹MT⁻¹ (base unit pascal second, kg · m⁻¹ · s⁻¹).

DynamicViscosity

Dynamic viscosity (base unit pascal second, kg · m⁻¹ · s⁻¹).