cu29_units::si

Module electric_charge

Electric charge (base unit coulomb, A · s).

Structs

abcoulomb

abcoulombs

ampere_hour

ampere hours

atomic_unit_of_charge

atomic units of charge

attocoulomb

attocoulombs

centicoulomb

centicoulombs

coulomb

Derived unit of electric charge.

decaampere_hour

decaampere hours

decacoulomb

decacoulombs

decicoulomb

decicoulombs

elementary_charge

Electric charge carried by a single proton.

exacoulomb

exacoulombs

faraday

faradays

femtocoulomb

femtocoulombs

franklin

franklins

gigaampere_hour

gigaampere hours

gigacoulomb

gigacoulombs

hectoampere_hour

hectoampere hours

hectocoulomb

hectocoulombs

kiloampere_hour

kiloampere hours

kilocoulomb

kilocoulombs

megaampere_hour

megaampere hours

megacoulomb

megacoulombs

microampere_hour

microampere hours

microcoulomb

microcoulombs

milliampere_hour

milliampere hours

millicoulomb

millcoulombs

nanocoulomb

nanocoulombs

petaampere_hour

petaampere hours

petacoulomb

petacoulombs

picocoulomb

picocoulombs

statcoulomb

statcoulombs

teraampere_hour

teraampere hours

teracoulomb

teracoulombs

yoctocoulomb

yoctocoulombs

yottacoulomb

yottacoulombs

zeptocoulomb

zeptocoulombs

zettacoulomb

zettacoulombs

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of electric charge, TI (base unit coulomb, A · s).

ElectricCharge

Electric charge (base unit coulomb, A · s).