cu29_units::si

Module electric_permittivity

Electric permittivity (base unit farad per meter, m⁻³ · kg⁻¹ · s⁴ · A²).

Structs

farad_per_meter

farads per meter

vacuum_electric_permittivity

vacuum electric permittivity

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of electric permittivity, L⁻³M⁻¹T⁴I² (base unit farad per meter, m⁻³ · kg⁻¹ · s⁴ · A²).

ElectricPermittivity

Electric permittivity (base unit farad per meter, m⁻³ · kg⁻¹ · s⁴ · A²).