cu29_units::si

Module electric_potential

Electric potential (base unit volt, m² · kg · s⁻³ · A⁻¹).

Structs

abvolt

abvolts

attovolt

attovolts

centivolt

centivolts

decavolt

decavolts

decivolt

decivolts

exavolt

exavolts

femtovolt

femtovolts

gigavolt

gigavolts

hectovolt

hectovolts

kilovolt

kilovolts

megavolt

megavolts

microvolt

microvolts

millivolt

millivolts

nanovolt

nanovolts

petavolt

petavolts

picovolt

picovolts

statvolt

statvolts

teravolt

teravolts

volt

Derived unit of electric potential.

yoctovolt

yoctovolts

yottavolt

yottavolts

zeptovolt

zeptovolts

zettavolt

zettavolts

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of electric potential, L²MT⁻³I⁻¹ (base unit volt, m² · kg · s⁻³ · A⁻¹).

ElectricPotential

Electric potential (base unit volt, m² · kg · s⁻³ · A⁻¹).