cu29_units::si

Module electric_quadrupole_moment

Electric quadrupole moment (base unit coulomb square meter, m² · s · A).

Structs

atomic_unit_of_electric_quadrupole_moment

Hartree unit of electric quadrupole moment e · a₀², where e is elementary charge and a₀ is Bohr radius.

coulomb_square_meter

coulomb square meters

elementary_charge_barn

elementary charge barns

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of electric quadrupole moment, LTI (base unit coulomb square meter, m² · s · A).

ElectricQuadrupoleMoment

Electric quadrupole moment (base unit coulomb square meter, m² · s · A).