cu29_units::si

Module electrical_resistance

Electrical resistance (base unit ohm, m² · kg · s⁻³ · A⁻²).

Structs

abohm

abohms

attoohm

attoohms

centiohm

centiohms

decaohm

decaohms

deciohm

deciohms

exaohm

exaohms

femtoohm

femtoohms

gigaohm

gigaohms

hectoohm

hectoohms

kiloohm

kiloohms

megaohm

megaohms

microohm

microohms

milliohm

milliohms

nanoohm

nanoohms

ohm

Derived unit of electrical resistance.

petaohm

petaohms

picoohm

picoohms

statohm

statohms

teraohm

teraohms

yoctoohm

yoctoohms

yottaohm

yottaohms

zeptoohm

zeptoohms

zettaohm

zettaohms

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of electrical resistance, L²MT⁻³I⁻² (base unit ohm, m² · kg · s⁻³ · A⁻²).

ElectricalResistance

Electrical resistance (base unit ohm, m² · kg · s⁻³ · A⁻²).