cu29_units::si

Module electrical_resistivity

Electrical resistivity (base unit ohm meter, m³ · kg · s⁻³ · A⁻²).

Structs

abohm_centimeter

abohm centimeters

abohm_meter

abohm meters

attoohm_meter

attoohm meters

centiohm_meter

centiohm meters

decaohm_meter

decaohm meters

deciohm_meter

deciohm meters

exaohm_meter

exaohm meters

femtoohm_meter

femtoohm meters

gigaohm_meter

gigaohm meters

hectoohm_meter

hectoohm meters

kiloohm_meter

kiloohm meters

megaohm_meter

megaohm meters

microohm_meter

microohm meters

milliohm_meter

milliohm meters

nanoohm_meter

nanoohm meters

ohm_centimeter

ohm centimeters

ohm_foot

ohm feet

ohm_inch

ohm inches

ohm_meter

Derived unit of electrical resistance.

ohm_square_millimeter_per_meter

ohm square millimeters per meter

ohm_yard

ohm yards

petaohm_meter

petaohm meters

picoohm_meter

picoohm meters

statohm_centimeter

statohm centimeters

statohm_meter

statohm meters

teraohm_meter

teraohm meters

yoctoohm_meter

yoctoohm meters

yottaohm_meter

yottaohm meters

zeptoohm_meter

zeptoohm meters

zettaohm_meter

zettaohm meters

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of electrical resistivity, L³MT⁻³I⁻² (base unit ohm meter, m³ · kg · s⁻³ · A⁻²).

ElectricalResistivity

Electrical resistivity (base unit ohm meter, m³ · kg · s⁻³ · A⁻²).