cu29_units::si

Module frequency

Frequency (base unit hertz, s⁻¹).

Structs

attohertz

attohertz

centihertz

centihertz

cycle_per_day

cycles per day

cycle_per_hour

cycles per hour

cycle_per_minute

cycles per minute

cycle_per_shake

cycles per shake

cycle_per_year

cycles per year

decahertz

decahertz

decihertz

decihertz

exahertz

exahertz

femtohertz

femtohertz

gigahertz

gigahertz

hectohertz

hectohertz

hertz

The hertz is one cycle per second.

kilohertz

kilohertz

megahertz

megahertz

microhertz

microhertz

millihertz

millihertz

nanohertz

nanohertz

petahertz

petahertz

picohertz

picohertz

terahertz

terahertz

yoctohertz

yoctohertz

yottahertz

yottahertz

zeptohertz

zeptohertz

zettahertz

zettahertz

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of frequency, T⁻¹ (base unit hertz, s⁻¹).

Frequency

Frequency (base unit hertz, s⁻¹).