cu29_units::si

Module heat_capacity

Heat capacity (base unit joule per kelvin, kg · m² · s⁻² · K⁻¹).

Heat capacity has the same kind as temperature interval, as this quantity relates to change of temperature. Not of kind TemperatureKind, used by thermodynamic temperature. See thermodynamic temperature for a full explanation.

This quantity might be used to define the heat capacity of an object. To define heat capacity of a material, use specific heat capacity.

Structs

attogram_square_meter_per_second_squared_kelvin

attogram square meters per second squared kelvin

attojoule_per_kelvin

attojoules per kelvin

boltzmann_constant

Boltzmann constants

btu_it_per_degree_fahrenheit

British thermal units (IT) per degree Fahrenheit

btu_per_degree_fahrenheit

British thermal units per degree Fahrenheit

calorie_per_kelvin

calories per kelvin

centigram_square_meter_per_second_squared_kelvin

centigram square meters per second squared kelvin

centijoule_per_kelvin

centijoules per kelvin

decagram_square_meter_per_second_squared_kelvin

decagram square meters per second squared kelvin

decajoule_per_kelvin

decajoules per kelvin

decigram_square_meter_per_second_squared_kelvin

decigram square meters per second squared kelvin

decijoule_per_kelvin

decijoules per kelvin

exagram_square_meter_per_second_squared_kelvin

exagram square meters per second squared kelvin

exajoule_per_kelvin

exajoules per kelvin

femtogram_square_meter_per_second_squared_kelvin

femtogram square meters per second squared kelvin

femtojoule_per_kelvin

femtojoules per kelvin

gigagram_square_meter_per_second_squared_kelvin

gigagram square meters per second squared kelvin

gigajoule_per_kelvin

gigajoules per kelvin

gram_square_meter_per_second_squared_kelvin

gram square meters per second squared kelvin

hectogram_square_meter_per_second_squared_kelvin

hectogram square meters per second squared kelvin

hectojoule_per_kelvin

hectojoules per kelvin

joule_per_degree_celsius

joules per degree celsius

joule_per_kelvin

Derived unit of heat capacity expressed in terms of derived unit Joule. Equivalent to kg · m²/(s² · K).

kilogram_square_meter_per_second_squared_kelvin

The derived unit of heat capacity expressed in base units. Equivalent to J/K.

kilojoule_per_degree_celsius

kilojoules per degree celsius

kilojoule_per_kelvin

kilojoules per kelvin

megagram_square_meter_per_second_squared_kelvin

megagram square meters per second squared kelvin

megajoule_per_kelvin

megajoules per kelvin

microgram_square_meter_per_second_squared_kelvin

microgram square meters per second squared kelvin

microjoule_per_kelvin

microjoules per kelvin

milligram_square_meter_per_second_squared_kelvin

milligram square meters per second squared kelvin

millijoule_per_degree_celsius

millijoules per degree celsius

millijoule_per_kelvin

millijoules per kelvin

nanogram_square_meter_per_second_squared_kelvin

nanogram square meters per second squared kelvin

nanojoule_per_kelvin

nanojoules per kelvin

petagram_square_meter_per_second_squared_kelvin

petagram square meters per second squared kelvin

petajoule_per_kelvin

petajoules per kelvin

picogram_square_meter_per_second_squared_kelvin

picogram square meters per second squared kelvin

picojoule_per_kelvin

picojoules per kelvin

teragram_square_meter_per_second_squared_kelvin

teragram square meters per second squared kelvin

terajoule_per_kelvin

terajoules per kelvin

yoctogram_square_meter_per_second_squared_kelvin

yoctogram square meters per second squared kelvin

yoctojoule_per_kelvin

yoctojoules per kelvin

yottagram_square_meter_per_second_squared_kelvin

yottagram square meters per second squared kelvin

yottajoule_per_kelvin

yottajoules per kelvin

zeptogram_square_meter_per_second_squared_kelvin

zeptogram square meters per second squared kelvin

zeptojoule_per_kelvin

zeptojoules per kelvin

zettagram_square_meter_per_second_squared_kelvin

zettagram square meters per second squared kelvin

zettajoule_per_kelvin

zettajoules per kelvin

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of heat capacity, L²MT⁻²Th⁻¹ (base unit joule per kelvin, kg · m² · s⁻² · K⁻¹).

HeatCapacity

Heat capacity (base unit joule per kelvin, kg · m² · s⁻² · K⁻¹).