cu29_units::si

Module heat_flux_density

Heat flux density (base unit watt per square meter, kg · s⁻³).

Structs

attowatt_per_square_meter

attowatts per square meter

centiwatt_per_square_meter

centiwatts per square meter

decawatt_per_square_meter

decawatts per square meter

deciwatt_per_square_meter

deciwatts per square meter

exawatt_per_square_meter

exawatts per square meter

femtowatt_per_square_meter

femtowatts per square meter

gigawatt_per_square_meter

gigawatts per square meter

hectowatt_per_square_meter

hectowatts per square meter

kilowatt_per_square_meter

Derived unit of density.

megawatt_per_square_meter

megawatts per square meter

microwatt_per_square_meter

microwatts per square meter

milliwatt_per_square_meter

milliwatts per square meter

nanowatt_per_square_meter

nanowatts per square meter

petawatt_per_square_meter

petawatts per square meter

picowatt_per_square_meter

picowatts per square meter

terawatt_per_square_meter

terawatts per square meter

watt_per_square_centimeter

watts per square centimeter

watt_per_square_meter

watts per square meter

watt_per_square_millimeter

watts per square millimeter

yoctowatt_per_square_meter

yoctowatts per square meter

yottawatt_per_square_meter

yottawatts per square meter

zeptowatt_per_square_meter

zeptowatts per square meter

zettawatt_per_square_meter

zettawatts per square meter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of heat flux density, MT⁻³ (base unit watt per square meter, kg · s⁻³).

HeatFluxDensity

Heat flux density (base unit watt per square meter, kg · s⁻³).