cu29_units::si

Module heat_transfer

Heat transfer (base unit watt per square meter kelvin, kg · s⁻³ · K⁻¹).

Heat transfer is also known as heat transfer coefficient, film coefficient, or film effectiveness. Commonly expressed using derived units power and area: watt per square meter kelvin.

Heat transfer has the same kind as temperature interval, as this quantity relates to change of temperature. Not of kind TemperatureKind, used by thermodynamic temperature. See thermodynamic temperature for a full explanation.

Structs

attogram_per_second_cubed_kelvin

attograms per second cubed kelvin

attowatt_per_square_meter_kelvin

attowatts per square meter kelvin

btu_it_per_hour_square_foot_degree_fahrenheit

British thermal units (IT) per hour square foot degree Fahrenheit

btu_it_per_hour_square_inch_degree_fahrenheit

British thermal units (IT) per hour square inch degree Fahrenheit

btu_it_per_minute_square_foot_degree_fahrenheit

British thermal units (IT) per minute square foot degree Fahrenheit

btu_it_per_minute_square_inch_degree_fahrenheit

British thermal units (IT) per minute square inch degree Fahrenheit

btu_it_per_second_square_foot_degree_fahrenheit

British thermal units (IT) per second square foot degree Fahrenheit

btu_it_per_second_square_inch_degree_fahrenheit

British thermal units (IT) per second square inch degree Fahrenheit

btu_per_hour_square_foot_degree_fahrenheit

British thermal units per hour square foot degree Fahrenheit

btu_per_hour_square_inch_degree_fahrenheit

British thermal units per hour square inch degree Fahrenheit

btu_per_minute_square_foot_degree_fahrenheit

British thermal units per minute square foot degree Fahrenheit

btu_per_minute_square_inch_degree_fahrenheit

British thermal units per minute square inch degree Fahrenheit

btu_per_second_square_foot_degree_fahrenheit

British thermal units per second square foot degree Fahrenheit

btu_per_second_square_inch_degree_fahrenheit

British thermal units per second square inch degree Fahrenheit

centigram_per_second_cubed_kelvin

centigrams per second cubed kelvin

centiwatt_per_square_meter_kelvin

centiwatts per square meter kelvin

decagram_per_second_cubed_kelvin

decagrams per second cubed kelvin

decawatt_per_square_meter_kelvin

decawatts per square meter kelvin

decigram_per_second_cubed_kelvin

decigrams per second cubed kelvin

deciwatt_per_square_meter_kelvin

deciwatts per square meter kelvin

exagram_per_second_cubed_kelvin

exagrams per second cubed kelvin

exawatt_per_square_meter_kelvin

exawatts per square meter kelvin

femtogram_per_second_cubed_kelvin

femtograms per second cubed kelvin

femtowatt_per_square_meter_kelvin

femtowatts per square meter kelvin

gigagram_per_second_cubed_kelvin

gigagrams per second cubed kelvin

gigawatt_per_square_meter_kelvin

gigawatts per square meter kelvin

gram_per_second_cubed_kelvin

grams per second cubed kelvin

hectogram_per_second_cubed_kelvin

hectograms per second cubed kelvin

hectowatt_per_square_meter_kelvin

hectowatts per square meter kelvin

joule_per_second_square_meter_degree_celsius

joules per second square meter degree celsius

joule_per_second_square_meter_kelvin

joules per second square meter kelvin

kilogram_per_second_cubed_degree_celsius

kilograms per second cubed degree celsius

kilogram_per_second_cubed_kelvin

Derived unit of heat transfer in base units. Equivalent to W/(m² · K).

kilowatt_per_square_meter_kelvin

kilowatts per square meter kelvin

megagram_per_second_cubed_kelvin

megagrams per second cubed kelvin

megawatt_per_square_meter_kelvin

megawatts per square meter kelvin

microgram_per_second_cubed_kelvin

micrograms per second cubed kelvin

microwatt_per_square_meter_kelvin

microwatts per square meter kelvin

milligram_per_second_cubed_kelvin

milligrams per second cubed kelvin

milliwatt_per_square_meter_kelvin

milliwatts per square meter kelvin

nanogram_per_second_cubed_kelvin

nanograms per second cubed kelvin

nanowatt_per_square_meter_kelvin

nanowatts per square meter kelvin

petagram_per_second_cubed_kelvin

petagrams per second cubed kelvin

petawatt_per_square_meter_kelvin

petawatts per square meter kelvin

picogram_per_second_cubed_kelvin

picograms per second cubed kelvin

picowatt_per_square_meter_kelvin

picowatts per square meter kelvin

teragram_per_second_cubed_kelvin

teragrams per second cubed kelvin

terawatt_per_square_meter_kelvin

terawatts per square meter kelvin

watt_per_square_centimeter_degree_celsius

watts per square centimeter degree celsius

watt_per_square_kilometer_degree_celsius

watts per square kilometer degree celsius

watt_per_square_meter_degree_celsius

watts per square meter degree celsius

watt_per_square_meter_kelvin

Derived unit of heat transfer in derived units. Equivalent to kg/(s³ · K).

watt_per_square_millimeter_degree_celsius

watts per square millimeter degree celsius

yoctogram_per_second_cubed_kelvin

yoctograms per second cubed kelvin

yoctowatt_per_square_meter_kelvin

yoctowatts per square meter kelvin

yottagram_per_second_cubed_kelvin

yottagrams per second cubed kelvin

yottawatt_per_square_meter_kelvin

yottawatts per square meter kelvin

zeptogram_per_second_cubed_kelvin

zeptograms per second cubed kelvin

zeptowatt_per_square_meter_kelvin

zeptowatts per square meter kelvin

zettagram_per_second_cubed_kelvin

zettagrams per second cubed kelvin

zettawatt_per_square_meter_kelvin

zettawatts per square meter kelvin

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of heat transfer, MT⁻³Th⁻¹ (base unit watt per square meter kelvin, kg · s⁻³ · K⁻¹).

HeatTransfer

Heat transfer (base unit watt per square meter kelvin, kg · s⁻³ · K⁻¹).