cu29_units::si

Module inductance

Inductance (base unit henry, m² · kg · s⁻² · A⁻²).

Structs

abhenry

abhenries

attohenry

attohenries

centihenry

centihenries

decahenry

decahenries

decihenry

decihenries

exahenry

exahenries

femtohenry

femtohenries

gigahenry

gigahenries

hectohenry

hectohenries

henry

Derived unit of inductance.

kilohenry

kilohenries

megahenry

megahenries

microhenry

microhenries

millihenry

millihenries

nanohenry

nanohenries

petahenry

petahenries

picohenry

picohenries

stathenry

stathenries

terahenry

terahenries

yoctohenry

yoctohenries

yottahenry

yottahenries

zeptohenry

zeptohenries

zettahenry

zettahenries

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of inductance, L²MT⁻²I⁻² (base unit henry, m² · kg · s⁻² · A⁻²).

Inductance

Inductance (base unit henry, m² · kg · s⁻² · A⁻²).