cu29_units::si

Module information_rate

Information rate (base unit byte per second, s⁻¹).

Structs

bit_per_second

bits per second

byte_per_second

bytes per second

exabit_per_second

exabits per second

exabyte_per_second

exabytes per second

exbibit_per_second

exbibits per second

exbibyte_per_second

exbibytes per second

gibibit_per_second

gibibits per second

gibibyte_per_second

gibibytes per second

gigabit_per_second

gigabits per second

gigabyte_per_second

gigabytes per second

kibibit_per_second

kibibits per second

kibibyte_per_second

kibibytes per second

kilobit_per_second

kilobits per second

kilobyte_per_second

kilobytes per second

mebibit_per_second

mebibits per second

mebibyte_per_second

mebibytes per second

megabit_per_second

megabits per second

megabyte_per_second

megabytes per second

octet_per_second

octets per second

pebibit_per_second

pebibits per second

pebibyte_per_second

pebibytes per second

petabit_per_second

petabits per second

petabyte_per_second

petabytes per second

tebibit_per_second

tebibits per second

tebibyte_per_second

tebibytes per second

terabit_per_second

terabits per second

terabyte_per_second

terabytes per second

yobibit_per_second

yobibits per second

yobibyte_per_second

yobibytes per second

yottabit_per_second

yottabits per second

yottabyte_per_second

yottabytes per second

zebibit_per_second

zebibits per second

zebibyte_per_second

zebibytes per second

zettabit_per_second

zettabits per second

zettabyte_per_second

zettabytes per second

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of information rate, T⁻¹ (base unit byte per second, s⁻¹).

InformationRate

Information rate (base unit byte per second, s⁻¹).