cu29_units::si

Module inverse_velocity

Inverse velocity (base unit second per meter, s · m⁻¹).

Structs

attosecond_per_meter

attoseconds per meter

centisecond_per_meter

centiseconds per meter

decasecond_per_meter

decaseconds per meter

decisecond_per_meter

deciseconds per meter

exasecond_per_meter

exaseconds per meter

femtosecond_per_meter

femtoseconds per meter

gigasecond_per_meter

gigaseconds per meter

hectosecond_per_meter

hectoseconds per meter

hour_per_foot

hours per foot

hour_per_kilometer

hours per kilometer

hour_per_mile

hours per mile

kilosecond_per_meter

kiloseconds per meter

megasecond_per_meter

megaseconds per meter

microsecond_per_meter

microseconds per meter

millisecond_per_meter

milliseconds per meter

minute_per_foot

minutes per foot

minute_per_kilometer

minutes per kilometer

minute_per_mile

minutes per mile

nanosecond_per_meter

nanoseconds per meter

petasecond_per_meter

petaseconds per meter

picosecond_per_meter

picoseconds per meter

second_per_foot

seconds per foot

second_per_inch

seconds per inch

second_per_meter

seconds per meter

second_per_mile

seconds per mile

terasecond_per_meter

teraseconds per meter

yoctosecond_per_meter

yoctoseconds per meter

yottasecond_per_meter

yottaseconds per meter

zeptosecond_per_meter

zeptoseconds per meter

zettasecond_per_meter

zettaseconds per meter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of inverse velocity, TL⁻¹ (base unit second per meter, s · m⁻¹).

InverseVelocity

Inverse velocity (base unit second per meter, s · m⁻¹).