cu29_units::si

Module jerk

Jerk (base unit meter per second cubed, m · s⁻³).

Structs

attometer_per_second_cubed

attometers per second cubed

centimeter_per_second_cubed

centimeters per second cubed

decameter_per_second_cubed

decameters per second cubed

decimeter_per_second_cubed

decimeters per second cubed

exameter_per_second_cubed

exameters per second cubed

femtometer_per_second_cubed

femtometers per second cubed

foot_per_second_cubed

feet per second cubed

gigameter_per_second_cubed

gigameters per second cubed

hectometer_per_second_cubed

hectometers per second cubed

inch_per_second_cubed

inches per second cubed

kilometer_per_minute_cubed

kilometers per minute cubed

kilometer_per_second_cubed

kilometers per second cubed

megameter_per_second_cubed

megameters per second cubed

meter_per_second_cubed

meters per second cubed

micrometer_per_second_cubed

micrometers per second cubed

millimeter_per_second_cubed

millimeters per second cubed

nanometer_per_second_cubed

nanometers per second cubed

petameter_per_second_cubed

petameters per second cubed

picometer_per_second_cubed

picometers per second cubed

terameter_per_second_cubed

terameters per second cubed

yoctometer_per_second_cubed

yoctometers per second cubed

yottameter_per_second_cubed

yottameters per second cubed

zeptometer_per_second_cubed

zeptometers per second cubed

zettameter_per_second_cubed

zettameters per second cubed

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of jerk, LT⁻³ (base unit meter per second cubed, m · s⁻³).

Jerk

Jerk (base unit meter per second cubed, m · s⁻³).