cu29_units::si

Module length

Length (base unit meter, m).

Structs

actus

acta

angstrom

ångströms

astronomical_unit

astronomical units

atomic_unit_of_length

Atomic unit of length (Bohr radius).

attometer

attometers

bohr_radius

bohr radiuses

centimeter

centimeters

chain

chains

cubitum

cubita

data_mile

data miles

decameter

decameters

decempeda

decempedae

decimeter

decimeters

digitus

digiti

exameter

exameters

fathom

fathoms

femtometer

femtometers

fermi

fermis

foot

feet

foot_survey

feet (U.S. survey)

gigameter

gigameters

gradus

gradus

hectometer

hectometers

inch

inches

kilometer

kilometers

leuga

leugae

light_year

light years

megameter

megameters

meter

The meter is the SI unit of length. It is defined by taking the fixed numerical value of the speed of light in vacuum c to be 299 792 458 when expressed in the unit m s⁻¹, where the second is defined in terms of the caesium frequency ∆νCs.

microinch

microinches

micrometer

micrometers

micron

microns

mil

mils

mile

miles

mile_survey

miles (U.S. survey)

mille_passus

mille passa

millimeter

millimeters

nanometer

nanometers

nautical_mile

nautical miles

palmipes

palmipedes

palmus

palmas

palmus_maior

palmas maior

parsec

parsecs

passus

passa

pes

pedes

petameter

petameters

pica_computer

picas (computer)

pica_printers

picas (printer’s)

picometer

picometers

point_computer

points (computer)

point_printers

points (printer’s)

rod

rods

stadium

stadia

terameter

terameters

uncia

unciae

yard

yards

yoctometer

yoctometers

yottameter

yottameters

zeptometer

zeptometers

zettameter

zettameters

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of length, L (base unit meter, m).

Length

Length (base unit meter, m).