cu29_units::si

Module linear_mass_density

Linear mass density (base unit kilogram per meter, m⁻¹ · kg).

Structs

gram_per_centimeter

grams per centimeter

gram_per_kilometer

grams per kilometer

kilogram_per_meter

kilograms per meter

ounce_per_foot

ounces per foot

ounce_per_inch

ounces per inch

pound_per_foot

pounds per foot

pound_per_inch

pounds per inch

pound_per_yard

pounds per yard

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of linear mass density, L⁻¹M (base unit kilogram per meter, m⁻¹ · kg).

LinearMassDensity

Linear mass density (base unit kilogram per meter, m⁻¹ · kg).