cu29_units::si

Module linear_number_density

Linear number density (base unit 1 per meter, m⁻¹).

Structs

per_centimeter

per centimeter

per_decimeter

per decimeter

per_foot

per foot

per_foot_survey

foot (U.S. survey)

per_inch

per inch

per_kilometer

per kilometer

per_meter

per meter

per_mile

per mile

per_mile_survey

per mile (U.S. survey)

per_millimeter

per millimeter

per_nautical_mile

per nautical mile

per_yard

per yard

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of linear number density, L⁻¹ (base 1 unit per meter, m⁻¹).

LinearNumberDensity

Linear number density (base unit 1 per meter, m⁻¹).