cu29_units::si

Module luminance

Luminance (base unit candela per square meter, cd · m⁻²).

Structs

attocandela_per_square_meter

attocandelas per square meter

candela_per_square_centimeter

candelas per square centimeter

candela_per_square_foot

candelas per square foot

candela_per_square_gigameter

candelas per square gigameter

candela_per_square_inch

candelas per square inch

candela_per_square_kilometer

candelas per square kilometer

candela_per_square_megameter

candelas per square megameter

candela_per_square_meter

Derived unit of luminance.

candela_per_square_micrometer

candelas per square micrometer

candela_per_square_millimeter

candelas per square millimeter

candela_per_square_nanometer

candelas per square nanometer

candela_per_square_picometer

candelas per square picometer

candela_per_square_terameter

candelas per square terameter

centicandela_per_square_meter

centicandelas per square meter

decacandela_per_square_meter

decacandelas per square meter

decicandela_per_square_meter

decicandelas per square meter

exacandela_per_square_meter

exacandelas per square meter

femtocandela_per_square_meter

femtocandelas per square meter

footlambert

footlamberts

gigacandela_per_square_meter

gigacandelas per square meter

hectocandela_per_square_meter

hectocandelas per square meter

kilocandela_per_square_meter

kilocandelas per square meter

lambert

lamberts

megacandela_per_square_meter

megacandelas per square meter

microcandela_per_square_meter

microcandelas per square meter

millicandela_per_square_meter

millicandelas per square meter

nanocandela_per_square_meter

nanocandelas per square meter

petacandela_per_square_meter

petacandelas per square meter

picocandela_per_square_meter

picocandelas per square meter

stilb

stilbs

teracandela_per_square_meter

teracandelas per square meter

yoctocandela_per_square_meter

yoctocandelas per square meter

yottacandela_per_square_meter

yottacandelas per square meter

zeptocandela_per_square_meter

zeptocandelas per square meter

zettacandela_per_square_meter

zettacandelas per square meter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of luminance, L⁻²J (base unit candela per square meter, cd · m⁻²).

Luminance

Luminance (base unit candela per square meter, cd · m⁻²).