cu29_units::si

Module mass

Mass (base unit kilogram, kg).

Structs

attogram

attograms

bes

bessis

carat

carats

centigram

centigrams

dalton

Unified atomic mass unit.

decagram

decagrams

decigram

decigrams

deunx

deunces

dextans

dextantes

dodrans

dodrantes

duella

duella

exagram

exagrams

femtogram

femtograms

gigagram

gigagrams

grain

grains

gram

grams

hectogram

hectograms

hundredweight_long

hundredweight (long)

hundredweight_short

hundredweight (short)

kilogram

The kilogram is the SI unit of mass. It is defined by taking the fixed numerical value of the Planck constant h to be 6.626 070 15 × 10⁻³⁴ when expressed in the unit J s, which is equal to kg m² s⁻¹, where the meter and the second are defined in terms of c and ∆νCs.

libra

librae

megagram

megagrams

microgram

micrograms

milligram

milligrams

nanogram

nanograms

obolus

oboli

ounce

ounces

ounce_troy

troy ounces

pennyweight

pennyweight

petagram

petagrams

picogram

picograms

pound

pounds

pound_troy

troy pounds

quadrans_teruncius

quadrantes teruncius

quincunx

quincunx

scrupulum

scrupula

semis

semisses

semisextula

semisextulae

semuncia

semunciae

septunx

septunces

sescuncia

sescunciae

sextans

sextantes

sextula

sextulae

sicilicus

scilici

siliqua

siliquae

slug

slugs

teragram

teragrams

ton

Ton Metric

ton_assay

assay tons

ton_long

long tons

ton_short

short tons

triens

trientes

uncia

unciae

yoctogram

yoctograms

yottagram

yottagrams

zeptogram

zeptograms

zettagram

zettagrams

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Mass dimension, M (base unit kilogram, kg).

Mass

Mass (base unit kilogram, kg).