cu29_units::si

Module mass_density

Mass density (base unit kilogram per cubic meter, kg · m⁻³).

Structs

attogram_per_cubic_meter

attograms per cubic meter

carat_per_cubic_meter

carats per cubic meter

centigram_per_cubic_meter

centigrams per cubic meter

decagram_per_cubic_meter

decagrams per cubic meter

decigram_per_cubic_meter

decigrams per cubic meter

exagram_per_cubic_meter

exagrams per cubic meter

femtogram_per_cubic_meter

femtograms per cubic meter

gigagram_per_cubic_meter

gigagrams per cubic meter

grain_per_cubic_meter

grains per cubic meter

grain_per_gallon

grains per gallon

gram_per_cubic_centimeter

grams per cubic centimeter

gram_per_cubic_meter

grams per cubic meter

hectogram_per_cubic_meter

hectograms per cubic meter

hundredweight_long_per_cubic_meter

hundredweight (long) per cubic meter

hundredweight_short_per_cubic_meter

hundredweight (short) per cubic meter

kilogram_per_cubic_meter

Derived unit of density.

megagram_per_cubic_meter

megagrams per cubic meter

microgram_per_cubic_meter

micrograms per cubic meter

milligram_per_cubic_meter

milligrams per cubic meter

nanogram_per_cubic_meter

nanograms per cubic meter

ounce_per_cubic_inch

ounces per cubic inch

ounce_per_cubic_meter

ounces per cubic meter

ounce_per_gallon

ounces per gallon

ounce_per_gallon_imperial

ounces per Imperial gallon

ounce_troy_per_cubic_meter

troy ounces per cubic meter

pennyweight_per_cubic_meter

pennyweight per cubic meter

petagram_per_cubic_meter

petagrams per cubic meter

picogram_per_cubic_meter

picograms per cubic meter

pound_per_cubic_foot

pounds per cubic foot

pound_per_cubic_inch

pounds per cubic inch

pound_per_cubic_meter

pounds per cubic meter

pound_per_cubic_yard

pounds per cubic yard

pound_per_gallon

pounds per gallon

pound_per_gallon_imperial

pounds per Imperial gallon

pound_troy_per_cubic_meter

troy pounds per cubic meter

slug_per_cubic_foot

slugs per cubic foot

slug_per_cubic_meter

slugs per cubic meter

teragram_per_cubic_meter

teragrams per cubic meter

ton_assay_per_cubic_meter

assay tons per cubic meter

ton_long_per_cubic_foot

long tons per cubic foot

ton_long_per_cubic_meter

long tons per cubic meter

ton_long_per_cubic_yard

long tons per cubic yard

ton_per_cubic_meter

tons per cubic meter

ton_short_per_cubic_foot

short tons per cubic foot

ton_short_per_cubic_meter

short tons per cubic meter

ton_short_per_cubic_yard

short tons per cubic yard

yoctogram_per_cubic_meter

yoctograms per cubic meter

yottagram_per_cubic_meter

yottagrams per cubic meter

zeptogram_per_cubic_meter

zeptograms per cubic meter

zettagram_per_cubic_meter

zettagrams per cubic meter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

DensityDeprecated

Mass density (base unit kilogram per cubic meter, kg · m⁻³).

Dimension

Dimension of mass density, L⁻³M (base unit kilogram per cubic meter, kg · m⁻³).

MassDensity

Mass density (base unit kilogram per cubic meter, kg · m⁻³).