cu29_units::si

Module mass_per_energy

Mass per energy (base unit kilogram per joule, m⁻² · s²).

This quantity is typically used to express an emission intensity, also known as carbon intensity. Emission intensity is a measure of how much mass of carbon dioxide (CO2) is emitted per unit of energy.

Structs

centigram_per_joule

centigrams per joule

decagram_per_joule

decagrams per joule

decigram_per_joule

decigrams per joule

gigagram_per_joule

gigagrams per joule

gram_per_gigawatt_hour

grams per gigawatt hour

gram_per_joule

grams per joule

gram_per_kilowatt_hour

grams per kilowatt hour

gram_per_megawatt_hour

grams per megawatt hour

gram_per_watt_hour

grams per watt hour

hectogram_per_joule

hectograms per joule

kilogram_per_gigawatt_hour

kilograms per gigawatt hour

kilogram_per_joule

Derived unit of Mass per energy.

kilogram_per_kilowatt_hour

kilograms per kilowatt hour

kilogram_per_megawatt_hour

kilograms per megawatt hour

kilogram_per_watt_hour

kilograms per watt hour

megagram_per_joule

megagrams per joule

microgram_per_joule

micrograms per joule

milligram_per_joule

milligrams per joule

pound_per_gigawatt_hour

pounds per gigawatt hour

pound_per_joule

pounds per joule

pound_per_kilowatt_hour

pounds per kilowatt hour

pound_per_megawatt_hour

pounds per megawatt hour

pound_per_watt_hour

pounds per watt hour

teragram_per_joule

teragrams per joule

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of mass per energy, L⁻²T² (base unit kilogram per joule, m⁻² · s²).

MassPerEnergy

Mass per energy (base unit kilogram per joule, m⁻² · s²).