cu29_units::si

Module molar_energy

Molar energy (base unit joule per mole, kg · m² · s⁻² · mol⁻¹).

Structs

attojoule_per_mole

attojoules per mole

btu_39_per_mole

British thermal units (39 °F) per mole

btu_59_per_mole

British thermal units (59 °F) per mole

btu_60_per_mole

British thermal units (60 °F) per mole

btu_it_per_mole

British thermal units (IT) per mole

btu_per_mole

British thermal units per mole

calorie_15_per_mole

calories (15 °C) per mole

calorie_20_per_mole

calories (20 °C) per mole

calorie_it_nutrition_per_mole

Calories (IT) per mole

calorie_it_per_mole

calories (IT) per mole

calorie_nutrition_per_mole

Calories per mole

calorie_per_mole

calories per mole

centijoule_per_mole

centijoules per mole

decajoule_per_mole

decajoules per mole

decawatt_hour_per_mole

decawatt hours per mole

decijoule_per_mole

decijoules per mole

electronvolt_per_mole

electronvolts per mole

electronvolt_per_particle

electronvolts per particle

erg_per_mole

ergs per mole

exajoule_per_mole

exajoules per mole

femtojoule_per_mole

femtojoules per mole

foot_pound_force_per_mole

foot pounds-force per mole

foot_poundal_per_mole

foot poundals per mole

gigajoule_per_mole

gigajoules per mole

gigawatt_hour_per_mole

gigawatt hours per mole

hectojoule_per_mole

hectojoules per mole

hectowatt_hour_per_mole

hectowatt hours per mole

joule_per_mole

The derived unit of molar energy. Equivalent to kg · m²/(s² · mol).

joule_per_particle

joules per particle

kilocalorie_it_per_mole

kilocalories (IT) per mole

kilocalorie_per_mole

kilocalories per mole

kilogram_square_meter_per_second_squared_mole

Molar energy expressed in base units. Equivalent to J/mol.

kilojoule_per_mole

kilojoules per mole

kilowatt_hour_per_mole

kilowatt hours per mole

megajoule_per_mole

megajoules per mole

megawatt_hour_per_mole

megawatt hours per mole

microjoule_per_mole

microjoules per mole

microwatt_hour_per_mole

microwatt hours per mole

millijoule_per_mole

millijoules per mole

milliwatt_hour_per_mole

milliwatt hours per mole

nanojoule_per_mole

nanojoules per mole

petajoule_per_mole

petajoules per mole

petawatt_hour_per_mole

petawatt hours per mole

picojoule_per_mole

picojoules per mole

quad_per_mole

quads per mole

terajoule_per_mole

terajoules per mole

terawatt_hour_per_mole

terawatt hours per mole

therm_ec_per_mole

therms (EC) per mole

therm_us_per_mole

therms per mole

ton_tnt_per_mole

tons of TNT per mole

watt_hour_per_mole

watt hours per mole

watt_second_per_mole

watt seconds per mole

yoctojoule_per_mole

yoctojoules per mole

yottajoule_per_mole

yottajoules per mole

zeptojoule_per_mole

zeptojoules per mole

zettajoule_per_mole

zettajoules per mole

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of molar energy, L²MT⁻²N⁻¹ (base unit joule per mole, kg · m² · s⁻² · mol⁻¹).

MolarEnergy

Molar energy (base unit joule per mole, kg · m² · s⁻² · mol⁻¹).