cu29_units::si

Module power_rate

Power rate (base unit watt per second, m² · kg · s⁻⁴).

Structs

attowatt_per_second

attowatts per second

centiwatt_per_second

centiwatts per second

decawatt_per_second

decawatts per second

deciwatt_per_second

deciwatts per second

erg_per_second_squared

ergs per second squared

exawatt_per_second

exawatts per second

femtowatt_per_second

femtowatts per second

foot_pound_per_second_squared

foot pounds-force per second squared

gigawatt_per_second

gigawatts per second

hectowatt_per_second

hectowatts per second

horsepower_per_second

horsepower per second

horsepower_per_second_boiler

horsepower per second (boiler)

horsepower_per_second_electric

horsepower per second (electric)

horsepower_per_second_imperial

horsepower per second (Imperial)

horsepower_per_second_metric

metric horsepower per second

hydraulic_horsepower_per_second

hydraulic horsepower per second

kilowatt_per_second

kilowatts per second

megawatt_per_second

megawatts per second

microwatt_per_second

microwatts per second

milliwatt_per_second

milliwatts per second

nanowatt_per_second

nanowatts per second

petawatt_per_second

petawatts per second

picowatt_per_second

picowatts per second

terawatt_per_second

terawatts per second

watt_per_second

Derived unit of power rate.

yoctowatt_per_second

yoctowatts per second

yottawatt_per_second

yottawatts per second

zeptowatt_per_second

zeptowatts per second

zettawatt_per_second

zettawatts per second

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of power rate, L²MT⁻⁴ (base unit watt per second, m² · kg · s⁻⁴).

PowerRate

Power rate (base unit watt per second, m² · kg · s⁻⁴).