cu29_units::si

Module pressure

Pressure (base unit pascal, kg · m⁻¹ · s⁻²).

Structs

atmosphere

atmospheres

atmosphere_technical

atmospheres (technical)

attopascal

attopascals

bar

bar

centimeter_of_mercury

centimeters of mercury

centimeter_of_mercury_0

centimeters of mercury (0 °C)

centimeter_of_water

centimeters of water

centimeter_of_water_4

centimeters of water (4 °C)

centipascal

centipascals

decapascal

decapascals

decipascal

decipascals

dyne_per_square_centimeter

dynes per square centimeter

exapascal

exapascals

femtopascal

femtopascals

foot_of_mercury

feet of mercury

foot_of_water

feet of water

foot_of_water_39_2

feet of water (39.2 °F)

gigapascal

gigapascals

gram_force_per_square_centimeter

grams-force per square centimeter

hectopascal

hectopascals

inch_of_mercury

inches of mercury

inch_of_mercury_32

inches of mercury (32 °F)

inch_of_mercury_60

inches of mercury (60 °F)

inch_of_water

inches of water

inch_of_water_60

inches of water (60 °F)

inch_of_water_39_2

inches of water (39.2 °F)

kilogram_force_per_square_centimeter

kilograms-force per square centimeter

kilogram_force_per_square_meter

kilograms-force per square meter

kilogram_force_per_square_millimeter

kilograms-force per square millimeter

kilopascal

kilopascals

kip_per_square_inch

kips per square inch

megapascal

megapascals

micropascal

micropascals

millibar

millibar

millimeter_of_mercury

millimeters of mercury

millimeter_of_water

millimeters of water

millipascal

millipascals

millitorr

millitorr

nanopascal

nanopascals

newton_per_square_millimeter

newtons per square millimeter

pascal

Derived unit of pressure.

petapascal

petapascals

picopascal

picopascals

pound_force_per_square_foot

pounds-force per square foot

pound_force_per_square_inch

pounds-force per square inch

poundal_per_square_foot

poundals per square foot

psi

pounds-force per square inch

terapascal

terapascals

torr

torr

yoctopascal

yoctopascals

yottapascal

yottapascals

zeptopascal

zeptopascals

zettapascal

zettapascals

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of pressure, L⁻¹MT⁻² (base unit pascal, kg · m⁻¹ · s⁻²).

Pressure

Pressure (base unit pascal, kg · m⁻¹ · s⁻²).