cu29_units::si

Module specific_heat_capacity

Specific heat capacity (base unit joule per kilogram kelvin, m² · s⁻² · K⁻¹).

This quantity might be used to define the heat capacity of a material. To define heat capacity of an object, use heat capacity.

Specific heat capacity has the same kind as temperature interval, as this quantity relates to change of temperature. Not of kind TemperatureKind, used by thermodynamic temperature. See thermodynamic temperature for a full explanation.

Structs

attojoule_per_kilogram_kelvin

attojoules per kilogram kelvin

btu_it_per_ounce_degree_fahrenheit

British thermal units (IT) per ounce degree Fahrenheit

btu_it_per_pound_degree_fahrenheit

British thermal units (IT) per pound degree Fahrenheit

btu_it_per_ton_degree_fahrenheit

British thermal units (IT) per ton degree Fahrenheit

btu_per_ounce_degree_fahrenheit

British thermal units per ounce degree Fahrenheit

btu_per_pound_degree_fahrenheit

British thermal units per pound degree Fahrenheit

btu_per_ton_degree_fahrenheit

British thermal units per ton degree Fahrenheit

calorie_per_gram_kelvin

calories per gram kelvin

calorie_per_kilogram_kelvin

calories per kilogram kelvin

centijoule_per_kilogram_kelvin

centijoules per kilogram kelvin

decajoule_per_kilogram_kelvin

decajoules per kilogram kelvin

decijoule_per_kilogram_kelvin

decijoules per kilogram kelvin

exajoule_per_kilogram_kelvin

exajoules per kilogram kelvin

femtojoule_per_kilogram_kelvin

femtojoules per kilogram kelvin

gigajoule_per_kilogram_kelvin

gigajoules per kilogram kelvin

hectojoule_per_kilogram_kelvin

hectojoules per kilogram kelvin

joule_per_gram_degree_celsius

joules per gram degree celsius

joule_per_kilogram_degree_celsius

joules per kilogram degree celsius

joule_per_kilogram_kelvin

Derived unit of specific heat capacity expressed in derived units. Equivalent to m²/(s² · K).

kilojoule_per_gram_degree_celsius

kilojoules per gram degree celsius

kilojoule_per_kilogram_degree_celsius

kilojoules per kilogram degree celsius

kilojoule_per_kilogram_kelvin

kilojoules per kilogram kelvin

megajoule_per_kilogram_kelvin

megajoules per kilogram kelvin

microjoule_per_kilogram_kelvin

microjoules per kilogram kelvin

millijoule_per_gram_degree_celsius

millijoules per gram degree celsius

millijoule_per_kilogram_degree_celsius

millijoules per kilogram degree celsius

millijoule_per_kilogram_kelvin

millijoules per kilogram kelvin

nanojoule_per_kilogram_kelvin

nanojoules per kilogram kelvin

petajoule_per_kilogram_kelvin

petajoules per kilogram kelvin

picojoule_per_kilogram_kelvin

picojoules per kilogram kelvin

square_centimeter_per_second_squared_kelvin

square centimeters per second squared kelvin

square_kilometer_per_second_squared_kelvin

square kilometers per second squared kelvin

square_meter_per_second_squared_kelvin

The derived unit of specific heat capacity expressed in base units. Equivalent to J/(kg · K).

square_micrometer_per_second_squared_kelvin

square micrometers per second squared kelvin

square_millimeter_per_second_squared_kelvin

square millimeters per second squared kelvin

terajoule_per_kilogram_kelvin

terajoules per kilogram kelvin

yoctojoule_per_kilogram_kelvin

yoctojoules per kilogram kelvin

yottajoule_per_kilogram_kelvin

yottajoules per kilogram kelvin

zeptojoule_per_kilogram_kelvin

zeptojoules per kilogram kelvin

zettajoule_per_kilogram_kelvin

zettajoules per kilogram kelvin

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of specific heat capacity, L²T⁻²Th⁻¹ (base unit joule per kilogram kelvin, m² · s⁻² · K⁻¹).

SpecificHeatCapacity

Specific heat capacity (base unit joule per kilogram kelvin, m² · s⁻² · K⁻¹).