cu29_units::si

Module specific_volume

Specific volume (base unit cubic meter per kilogram, m³ · kg⁻¹).

Structs

cubic_centimeter_per_gram

cubic centimeters per gram

cubic_centimeter_per_kilogram

cubic centimeters per kilogram

cubic_foot_per_pound

cubic feet per pound

cubic_foot_per_slug

cubic feett per slug

cubic_foot_per_ton_long

cubic feet per long ton

cubic_foot_per_ton_short

cubic feet per short ton

cubic_inch_per_ounce

cubic inches per ounce

cubic_inch_per_pound

cubic inches per pound

cubic_meter_per_attogram

cubic meters per attogram

cubic_meter_per_carat

cubic meters per carat

cubic_meter_per_centigram

cubic meters per centigram

cubic_meter_per_decagram

cubic meters per decagram

cubic_meter_per_decigram

cubic meters per decigram

cubic_meter_per_exagram

cubic meters per exagram

cubic_meter_per_femtogram

cubic meters per femtogram

cubic_meter_per_gigagram

cubic meters per gigagram

cubic_meter_per_grain

cubic meters per grain

cubic_meter_per_gram

cubic meters per gram

cubic_meter_per_hectogram

cubic meters per hectogram

cubic_meter_per_hundredweight_long

cubic meters per hundredweight (long)

cubic_meter_per_hundredweight_short

cubic meters per hundredweight (short)

cubic_meter_per_kilogram

Derived unit of specific volume.

cubic_meter_per_megagram

cubic meters per megagram

cubic_meter_per_microgram

cubic meters per microgram

cubic_meter_per_milligram

cubic meters per milligram

cubic_meter_per_nanogram

cubic meters per nanogram

cubic_meter_per_ounce

cubic meters per ounce

cubic_meter_per_ounce_troy

cubic meters per troy ounce

cubic_meter_per_pennyweight

cubic meters per pennyweight

cubic_meter_per_petagram

cubic meters per petagram

cubic_meter_per_picogram

cubic meters per picogram

cubic_meter_per_pound

cubic meters per pound

cubic_meter_per_pound_troy

cubic meters per troy pound

cubic_meter_per_slug

cubic meters per slug

cubic_meter_per_teragram

cubic meters per teragram

cubic_meter_per_ton

cubic meters per ton

cubic_meter_per_ton_assay

cubic meters per assay ton

cubic_meter_per_ton_long

cubic meters per long ton

cubic_meter_per_ton_short

cubic meters per short ton

cubic_meter_per_yoctogram

cubic meters per yoctogram

cubic_meter_per_yottagram

cubic meters per yottagram

cubic_meter_per_zeptogram

cubic meters per zeptogram

cubic_meter_per_zettagram

cubic meters per zettagram

cubic_yard_per_pound

cubic yards per pound

cubic_yard_per_ton_long

cubic yards per long ton

cubic_yard_per_ton_short

cubic yards per short ton

gallon_imperial_per_ounce

Imperial gallons per ounce

gallon_imperial_per_pound

Imperial gallons per pound

gallon_per_grain

gallons per grain

gallon_per_ounce

gallons per ounce

gallon_per_pound

gallons per pound

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of specific volume, L³M⁻¹ (base unit cubic meter per kilogram, m³ · kg⁻¹).

SpecificVolume

Specific volume (base unit cubic meter per kilogram, m³ · kg⁻¹).