cu29_units::si

Module surface_tension

Surface tension (base unit newton per meter, kg · s⁻²).

Structs

attonewton_per_meter

attonewtons per meter

centinewton_per_meter

centinewtons per meter

decanewton_per_meter

decanewtons per meter

decinewton_per_meter

decinewtons per meter

dyne_per_centimeter

dynes per centimeter

exanewton_per_meter

exanewtons per meter

femtonewton_per_meter

femtonewtons per meter

giganewton_per_meter

giganewtons per meter

gram_force_per_centimeter

grams-force per centimeter

hectonewton_per_meter

hectonewtons per meter

joule_per_square_centimeter

joules per square centimeter

joule_per_square_meter

joules per square meter

kilogram_force_per_meter

kilograms-force per meter

kilonewton_per_meter

kilonewtons per meter

meganewton_per_meter

meganewtons per meter

micronewton_per_meter

micronewtons per meter

millinewton_per_meter

millinewtons per meter

nanonewton_per_meter

nanonewtons per meter

newton_per_meter

Derived unit of surface tension.

ounce_force_per_inch

ounces-force per inch

petanewton_per_meter

petanewtons per meter

piconewton_per_meter

piconewtons per meter

pound_force_per_inch

pounds-force per inch

poundal_per_inch

poundals per inch

teranewton_per_meter

teranewtons per meter

yoctonewton_per_meter

yoctonewtons per meter

yottanewton_per_meter

yottanewtons per meter

zeptonewton_per_meter

zeptonewtons per meter

zettanewton_per_meter

zettanewtons per meter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of surface tension, MT⁻² (base unit newton per meter, kg · s⁻²).

SurfaceTension

Surface tension (base unit newton per meter, kg · s⁻²).