cu29_units::si

Module time

Time (base unit second, s).

Structs

attosecond

attoseconds

centisecond

centiseconds

day

days

day_sidereal

days (sidereal)

decasecond

decaseconds

decisecond

deciseconds

exasecond

exaseconds

femtosecond

femtoseconds

gigasecond

gigaseconds

hectosecond

hectoseconds

hour

hours

hour_sidereal

hours (sidereal)

kilosecond

kiloseconds

megasecond

megaseconds

microsecond

microseconds

millisecond

milliseconds

minute

minutes

nanosecond

nanoseconds

petasecond

petaseconds

picosecond

picoseconds

second

The second is the SI unit of time. It is defined by taking the fixed numerical value of the caesium frequency ∆νCs, the unperturbed ground-state hyperfine transition frequency of the caesium 133 atom, to be 9 192 631 770 when expressed in the unit Hz, which is equal to s⁻¹.

second_sidereal

seconds (sidereal)

shake

shakes

terasecond

teraseconds

year

years

year_sidereal

years (sidereal)

year_tropical

years (tropical)

yoctosecond

yoctoseconds

yottasecond

yottaseconds

zeptosecond

zeptoseconds

zettasecond

zettaseconds

Enums

TryFromError

An error encountered converting between Time and Duration.

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of time, T (base unit second, s).

Time

Time (base unit second, s).