cu29_units::si

Module torque

Torque (aka moment of force) (base unit newton meter, kg · m² · s⁻²).

Structs

attonewton_meter

attonewton meters

centinewton_meter

centinewton meters

decanewton_meter

decanewton meters

decinewton_meter

decinewton meters

dyne_centimeter

dyne centimeters

dyne_meter

dyne meters

exanewton_meter

exanewton meters

femtonewton_meter

femtonewton meters

giganewton_meter

giganewton meters

hectonewton_meter

hectonewton meters

kilogram_force_meter

kilogram-force meters

kilonewton_meter

kilonewton meters

meganewton_meter

meganewton meters

micronewton_meter

micronewton meters

millinewton_meter

millinewton meters

nanonewton_meter

nanonewton meters

newton_attometer

newton attometers

newton_centimeter

newton centimeters

newton_decameter

newton decameters

newton_decimeter

newton decimeters

newton_exameter

newton exameters

newton_femtometer

newton femtometers

newton_gigameter

newton gigameters

newton_hectometer

newton hectometers

newton_kilometer

newton kilometers

newton_megameter

newton megameters

newton_meter

Derived unit of torque.

newton_micrometer

newton micrometers

newton_millimeter

newton millimeters

newton_nanometer

newton nanometers

newton_petameter

newton petameters

newton_picometer

newton picometers

newton_terameter

newton terameters

newton_yoctometer

newton yoctometers

newton_yottameter

newton yottameters

newton_zeptometer

newton zeptometers

newton_zettameter

newton zettameters

ounce_force_inch

ounces-force inches

petanewton_meter

petanewton meters

piconewton_meter

piconewton meters

pound_force_foot

pounds-force feet

pound_force_inch

pounds-force inches

teranewton_meter

teranewton meters

yoctonewton_meter

yoctonewton meters

yottanewton_meter

yottanewton meters

zeptonewton_meter

zeptonewton meters

zettanewton_meter

zettanewton meters

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of torque, L²MT⁻² (base unit newton meter, kg · m² · s⁻²).

Torque

Torque (aka moment of force) (base unit newton meter, kg · m² · s⁻²).