cu29_units::si

Module volumetric_power_density

Volumetric power density (base unit watt per cubic meter, m⁻¹ · kg · s⁻³).

Structs

watt_per_cubic_centimeter

watts per cubic centimeter

watt_per_cubic_meter

watts per cubic meter

watt_per_cubic_millimeter

watts per cubic millimeter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See [crate::Conversion<V>].

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of volumetric power density, L⁻¹MT⁻³ (base unit watt per cubic meter, m⁻¹ · kg · s⁻³).

VolumetricPowerDensity

Volumetric power density (base unit watt per cubic meter, m⁻¹ · kg · s⁻³).